makefile: Fix dependency output generation
To allow for compilable files in subdirectories,
the commit 05b9503 used the option -cd to latexmk.
However, latexmk generate the dependencies
relative to the new path, not to the original path.

The solution is to remove the -cd option
and use -outdir and -auxdir instead.
Note we save both the binary and the auxiliary files
in the subdirectory, for consistency with the other files.